ClienteFinal 2.0
en esta nueva version 2.0 se busca una mejora singificativa para el microservicio ya que es mucho mas eficiente al hacer las pruebas al ocupar dto como interseptores de dato y se simplifico el proyecto ya que avian cosas que  no solicitaban en el caso

+++ src/main/java/com/digipymes360/clienteFinal/repository/SoporteRepository.java
@@ -4,5 +4,6 @@ package com.digipymes360.clienteFinal.repository;
 import com.digipymes360.clienteFinal.model.Soporte;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SoporteRepository extends JpaRepository<Soporte, Integer> {
+public interface SoporteRepository extends JpaRepository<Soporte, Long> {
 }
+
